Fix clippy warnings workspace-wide
- Fix 3 genuine logic bugs found by clippy (always-true/false
  conditions in tests, an erasing multiplication) and one MutexGuard
  held across an await point in terminal_thread.rs.
- Simplify an Option<Option<T>> to Option<T> in the dialing directory
  double-click tracking.
- Apply hundreds of mechanical fixes: clone_from, Shadow::default(),
  redundant struct updates, lazy_static! -> std::sync::LazyLock,
  strip_prefix, needless range loops, manual Iterator::find,
  let...else -> ?, merged identical match arms, collapsed matches,
  Debug -> Display formatting in logs, removed dead imports/items.
- Remove the now-unused lazy_static dependency from icy_term and
  icy_draw.
- Add scoped, justified #[allow] for pedantic lints that would need
  broad public API/signature changes across many GUI call sites.

cargo clippy --workspace --all-targets is now warning-free except for
3 pre-existing dead-code warnings (unrelated to clippy, kept as they
may back future features). Full workspace test suite passes.

+++ crates/icy_engine_edit/tests/bitfont/font_operations/insert_delete_column.rs
@@ -38,7 +38,7 @@ fn test_insert_column_maximum_width() {
 
     // Set a pixel at rightmost column
     state.clear_glyph('A').unwrap();
-    state.set_pixel('A', (MAX_FONT_WIDTH - 1) as i32, 0, true).unwrap();
+    state.set_pixel('A', MAX_FONT_WIDTH - 1, 0, true).unwrap();
 
     // Try to insert column - should have no effect
     state.set_cursor_pos(0, 0);
